member refactor
Former-commit-id: 3c0a0fb3d475cc0103a81ae9a6594c1f865ba5fa

+++ core/domain/interface/after-sales/after_sales.go
@@ -175,7 +175,7 @@ type (
 		// 运营商编号
 		VendorId int32 `db:"vendor_id"`
 		// 购买者编号
-		BuyerId int32 `db:"buyer_id"`
+		BuyerId int64 `db:"buyer_id"`
 		// 类型，退货、换货、维修
 		Type int `db:"type"`
 		// 退货的商品项编号
